[CL 2956455 by Yannick Lange in Dev-VREditor branch]

+++ Engine/Source/Programs/AutomationTool/AutomationUtils/ProcessUtils.cs
@@ -534,8 +534,11 @@ namespace AutomationTool
 						}
 						catch (Exception Ex)
 						{
-							CommandUtils.LogWarning("Failed to kill descendant:");
-							CommandUtils.LogWarning(LogUtils.FormatException(Ex));
+							if(!KillCandidate.HasExited)
+							{
+								CommandUtils.LogWarning("Failed to kill descendant:");
+								CommandUtils.LogWarning(LogUtils.FormatException(Ex));
+							}
 						}
 						break;  // exit the loop as who knows what else died, so let's get processes anew
 					}
